refactor(PAY-1-E): split the remaining 13 over-budget files — check:architecture EXIT 0
Leaf extractions per assumption A-0001 inventory (largest: openapi-session-
schemas 203, tool-rule-runtime-command 157, agent-delegate-route-helpers
139, desired-state-capability-validation 142). All targets under budget;
typecheck green; 443 focused tests green; architecture checker passes
repo-wide for the first time.
