+++ apps/core/src/control/server/openapi-routes-extended.ts
@@ -1,4 +1,5 @@
 import { doc, ids, query, type RouteDoc } from './openapi-route-helpers.js';
+import { mcpOpenApiRouteDocs } from './openapi-mcp-routes.js';
 
 export const extendedOpenApiRouteDocs: RouteDoc[] = [
   doc(
@@ -630,107 +631,5 @@ export const extendedOpenApiRouteDocs: RouteDoc[] = [
     ['skills:admin'],
     { body: 'none', parameters: [ids.agent, ids.skill] },
   ),
-  doc(
-    'get',
-    '/v1/mcp-servers',
-    'listMcpServers',
-    'MCP Servers',
-    'List MCP servers',
-    'Lists MCP server definitions, optionally filtered by status.',
-    ['mcp:read'],
-    {
-      parameters: [
-        query('status', 'Server status filter.', {
-          type: 'string',
-          enum: ['active', 'disabled'],
-        }),
-        query('limit', 'Maximum number of servers.', {
-          type: 'integer',
-          minimum: 1,
-          maximum: 500,
-        }),
-        query('cursor', 'Pagination cursor.'),
-      ],
-    },
-  ),
-  doc(
-    'post',
-    '/v1/mcp-servers',
-    'connectMcpServer',
-    'MCP Servers',
-    'Connect MCP server',
-    'Connects and validates a current MCP server definition.',
-    ['mcp:admin'],
-    { body: 'json', status: '201' },
-  ),
-  doc(
-    'get',
-    '/v1/mcp-servers/{serverId}',
-    'getMcpServer',
-    'MCP Servers',
-    'Get MCP server',
-    'Returns one MCP server definition.',
-    ['mcp:read'],
-    { parameters: [ids.server] },
-  ),
-  doc(
-    'post',
-    '/v1/mcp-servers/{serverId}/disable',
-    'disableMcpServer',
-    'MCP Servers',
-    'Disable MCP server',
-    'Disables an active MCP server and syncs settings.',
-    ['mcp:admin'],
-    { body: 'json', parameters: [ids.server] },
-  ),
-  doc(
-    'post',
-    '/v1/mcp-servers/{serverId}/test',
-    'testMcpServer',
-    'MCP Servers',
-    'Test MCP server',
-    'Runs server-side MCP definition validation.',
-    ['mcp:admin'],
-    { body: 'json', parameters: [ids.server] },
-  ),
-  doc(
-    'get',
-    '/v1/agents/{agentId}/mcp-servers',
-    'listAgentMcpServerBindings',
-    'MCP Servers',
-    'List agent MCP server bindings',
-    'Lists MCP servers currently bound to an agent.',
-    ['mcp:read'],
-    { parameters: [ids.agent] },
-  ),
-  doc(
-    'put',
-    '/v1/agents/{agentId}/mcp-servers/{serverId}',
-    'bindMcpServerToAgent',
-    'MCP Servers',
-    'Bind MCP server to agent',
-    'Binds an active MCP server and syncs settings.',
-    ['mcp:admin', 'agents:admin'],
-    { body: 'json', parameters: [ids.agent, ids.server] },
-  ),
-  doc(
-    'patch',
-    '/v1/agents/{agentId}/mcp-servers/{serverId}',
-    'updateAgentMcpServerBinding',
-    'MCP Servers',
-    'Update agent MCP server binding',
-    'Updates binding policy metadata.',
-    ['mcp:admin', 'agents:admin'],
-    { body: 'json', parameters: [ids.agent, ids.server] },
-  ),
-  doc(
-    'delete',
-    '/v1/agents/{agentId}/mcp-servers/{serverId}',
-    'unbindMcpServerFromAgent',
-    'MCP Servers',
-    'Unbind MCP server from agent',
-    'Removes an MCP server binding and syncs settings.',
-    ['mcp:admin', 'agents:admin'],
-    { body: 'none', parameters: [ids.agent, ids.server] },
-  ),
+  ...mcpOpenApiRouteDocs,
 ];
